+++ entry/tests/test_api_v2.py
@@ -10,7 +10,6 @@ import yaml
 from rest_framework import status
 from rest_framework.exceptions import ValidationError
 
-from airone.lib.elasticsearch import FilterKey
 from airone.lib.log import Logger
 from airone.lib.test import AironeViewTest
 from airone.lib.types import (
@@ -36,27 +35,9 @@ from trigger.models import TriggerCondition
 from user.models import User
 
 
-class ViewTest(AironeViewTest):
-    def _create_user(
-        self,
-        name,
-        email="email",
-        is_superuser=False,
-        authenticate_type=User.AuthenticateType.AUTH_TYPE_LOCAL,
-    ):
-        user = User(
-            username=name,
-            email=email,
-            is_superuser=is_superuser,
-            authenticate_type=authenticate_type,
-        )
-        user.set_password(name)
-        user.save()
-
-        return user
-
+class BaseViewTest(AironeViewTest):
     def setUp(self):
-        super(ViewTest, self).setUp()
+        super(BaseViewTest, self).setUp()
 
         self.user: User = self.guest_login()
 
@@ -86,6 +67,26 @@ class ViewTest(AironeViewTest):
                 "attrs": attrs,
             }
         )
+
+
+class ViewTest(BaseViewTest):
+    def _create_user(
+        self,
+        name,
+        email="email",
+        is_superuser=False,
+        authenticate_type=User.AuthenticateType.AUTH_TYPE_LOCAL,
+    ):
+        user = User(
+            username=name,
+            email=email,
+            is_superuser=is_superuser,
+            authenticate_type=authenticate_type,
+        )
+        user.set_password(name)
+        user.save()
+
+        return user
 
     def test_retrieve_entry(self):
         entry: Entry = self.add_entry(
@@ -2161,7 +2162,7 @@ class ViewTest(AironeViewTest):
             {"result": "Succeed in registering export processing. Please check Job list."},
         )
 
-        job = Job.objects.last()
+        job = Job.objects.filter(target=entity).last()
         self.assertEqual(job.operation, JobOperation.EXPORT_ENTRY_V2)
         self.assertEqual(job.status, JobStatus.DONE)
 
@@ -2303,7 +2304,7 @@ class ViewTest(AironeViewTest):
             )
             self.assertEqual(resp.status_code, 200)
 
-            content = Job.objects.last().get_cache()
+            content = Job.objects.filter(target=test_entity).last().get_cache()
             header = content.splitlines()[0]
             self.assertEqual(header, 'Name,"%s,""ATTR"""' % type_name)
 
@@ -2989,6 +2990,7 @@ class ViewTest(AironeViewTest):
             resp.json(),
             {
                 "count": 2,
+                "total_count": 2,
                 "values": [
                     {
                         "entity": {"id": self.entity.id, "name": "test-entity"},
@@ -3154,189 +3156,6 @@ class ViewTest(AironeViewTest):
                 ],
             },
         )
-
-    def test_advanced_search_with_join_attrs(self):
-        # create Items to be search by join_attrs parameter
-        ref_entries = [
-            self.add_entry(
-                self.user,
-                "RefEntry-%s" % i,
-                self.ref_entity,
-                values={
-                    "val": "hoge-%s" % i if i % 2 == 0 else "",
-                    "ref": self.ref_entry.id,
-                    "name": {"name": "abcd-%s" % i, "id": self.ref_entry.id},
-                },
-            )
-            for i in range(2)
-        ]
-
-        # create Items that are search by ordinary processing
-        for index, (val, ref_id) in enumerate(
-            [
-                ("foo", ref_entries[0].id),
-                ("bar", ref_entries[0].id),
-                ("baz", ref_entries[1].id),
-                ("qux", None),
-            ]
-        ):
-            self.add_entry(
-                self.user,
-                "Entry%s" % index,
-                self.entity,
-                values={
-                    "val": val,
-                    "ref": ref_id,
-                    "name": {"name": "fuga", "id": ref_id},
-                },
-            )
-
-        # send request to search Entries with join_attrs
-        params = {
-            "entities": [self.entity.id],
-            "attrinfo": [
-                {"name": "val"},
-                {"name": "ref"},
-                {"name": "name"},
-            ],
-            "join_attrs": [
-                {
-                    "name": "ref",
-                    "attrinfo": [
-                        {"name": "val"},
-                        {"name": "ref"},
-                        {"name": "name"},
-                    ],
-                },
-            ],
-        }
-        resp = self.client.post(
-            "/entry/api/v2/advanced_search/", json.dumps(params), "application/json"
-        )
-        self.assertEqual(resp.status_code, 200)
-
-        # prepare comparison data with returned result
-        REF_DATA = {"id": self.ref_entry.id, "name": self.ref_entry.name}
-        expected_results = [
-            (
-                "Entry0",
-                {
-                    "ref.val": {"as_string": "hoge-0"},
-                    "ref.ref": {"as_object": REF_DATA},
-                    "ref.name": {"as_named_object": {"name": "abcd-0", "object": REF_DATA}},
-                },
-            ),
-            (
-                "Entry1",
-                {
-                    "ref.val": {"as_string": "hoge-0"},
-                    "ref.ref": {"as_object": REF_DATA},
-                    "ref.name": {"as_named_object": {"name": "abcd-0", "object": REF_DATA}},
-                },
-            ),
-            (
-                "Entry2",
-                {
-                    "ref.val": {"as_string": ""},
-                    "ref.ref": {"as_object": REF_DATA},
-                    "ref.name": {"as_named_object": {"name": "abcd-1", "object": REF_DATA}},
-                },
-            ),
-            (
-                "Entry3",
-                {
-                    "ref.val": {"as_string": ""},
-                    "ref.ref": {"as_string": ""},
-                    "ref.name": {"as_string": ""},
-                },
-            ),
-        ]
-
-        # check returned processing has expected values
-        for index, result in enumerate(resp.json()["values"]):
-            (e_name, e_attrinfo) = expected_results[index]
-            self.assertEqual(result["entry"]["name"], e_name)
-
-            for attrname, attrvalue in e_attrinfo.items():
-                self.assertEqual(result["attrs"][attrname]["value"], attrvalue)
-
-        # This sends request with keyword in join_attrs parameter and
-        # confirms filter processing works with its filter parameter of attrinfo
-        params = {
-            "entities": [self.entity.id],
-            "attrinfo": [
-                {"name": "val", "keyword": "bar"},
-                {"name": "ref"},
-            ],
-            "join_attrs": [
-                {
-                    "name": "ref",
-                    "attrinfo": [
-                        {"name": "val", "keyword": "hoge-0"},
-                    ],
-                },
-            ],
-        }
-        resp = self.client.post(
-            "/entry/api/v2/advanced_search/", json.dumps(params), "application/json"
-        )
-        self.assertEqual(resp.status_code, 200)
-        self.assertEqual([x["entry"]["name"] for x in resp.json()["values"]], ["Entry1"])
-
-        # This sends request with join_attrs that have filter_key to get empty Items
-        params = {
-            "entities": [self.entity.id],
-            "attrinfo": [],
-            "join_attrs": [
-                {
-                    "name": "ref",
-                    "attrinfo": [
-                        {"name": "val", "filter_key": FilterKey.EMPTY},
-                    ],
-                },
-            ],
-        }
-        resp = self.client.post(
-            "/entry/api/v2/advanced_search/", json.dumps(params), "application/json"
-        )
-        self.assertEqual(resp.status_code, 200)
-        self.assertEqual([x["entry"]["name"] for x in resp.json()["values"]], ["Entry2"])
-        self.assertEqual(
-            [x["attrs"]["ref.val"]["value"] for x in resp.json()["values"]], [{"as_string": ""}]
-        )
-
-        # This sends request with join_attrs that have filter_key to get non-empty Items
-        params = {
-            "entities": [self.entity.id],
-            "attrinfo": [],
-            "join_attrs": [
-                {
-                    "name": "ref",
-                    "attrinfo": [
-                        {"name": "val", "filter_key": FilterKey.NON_EMPTY},
-                    ],
-                },
-            ],
-        }
-        resp = self.client.post(
-            "/entry/api/v2/advanced_search/", json.dumps(params), "application/json"
-        )
-        self.assertEqual(resp.status_code, 200)
-        self.assertEqual([x["entry"]["name"] for x in resp.json()["values"]], ["Entry0", "Entry1"])
-        self.assertEqual(
-            [x["attrs"]["ref.val"]["value"] for x in resp.json()["values"]],
-            [
-                {"as_string": "hoge-0"},
-                {"as_string": "hoge-0"},
-            ],
-        )
-
-    def test_advanced_search_with_join_attrs_that_have_filter_key1(self):
-        """
-        Test advanced_search APIv2 with join_attrs parameter that have
-        valid filter_key attribute.
-        """
-        pass
 
     def test_advanced_search_all_entities(self):
         params = {
@@ -4316,7 +4135,7 @@ class ViewTest(AironeViewTest):
         referrals = referred_entity["entries"][0]["referrals"]
         self.assertEqual(len(referrals), 1)
         self.assertEqual(referrals[0]["entity"], "entity")
-        self.assertEqual(referrals[0]["entry"], "entry")
+        self.assertEqual(referrals[0]["entries"], "entry")
 
     @patch(
         "entry.tasks.export_search_result_v2.delay", Mock(side_effect=tasks.export_search_result_v2)
@@ -4457,6 +4276,76 @@ class ViewTest(AironeViewTest):
                     "entries": [{"attrs": [{"name": "val", "value": "hoge"}], "name": "Entry"}],
                 },
             ],
+        )
+
+    def test_advanced_search_chain(self):
+        ref_entry = self.add_entry(self.user, "RefEntry", self.ref_entity, values={"val": "hoge"})
+        entry = self.add_entry(
+            self.user,
+            "Entry",
+            self.entity,
+            values={
+                "ref": ref_entry.id,
+            },
+        )
+
+        params = {
+            "entities": [self.entity.id],
+            "attrs": [{"name": "ref", "attrs": [{"name": "val", "value": "hoge"}]}],
+        }
+        resp = self.client.post(
+            "/entry/api/v2/advanced_search_chain/", json.dumps(params), "application/json"
+        )
+        self.assertEqual(resp.status_code, 200)
+        self.assertEqual(
+            resp.json(),
+            [
+                {
+                    "id": entry.id,
+                    "name": "Entry",
+                    "schema": {
+                        "id": entry.schema.id,
+                        "name": "test-entity",
+                        "is_public": True,
+                    },
+                    "is_active": True,
+                    "deleted_user": None,
+                    "deleted_time": None,
+                    "updated_time": entry.updated_time.astimezone(self.TZ_INFO).isoformat(),
+                },
+            ],
+        )
+
+        # empty result case
+        params = {
+            "entities": [self.entity.id],
+            "attrs": [{"name": "ref", "attrs": [{"name": "val", "value": "fuga"}]}],
+        }
+        resp = self.client.post(
+            "/entry/api/v2/advanced_search_chain/", json.dumps(params), "application/json"
+        )
+        self.assertEqual(resp.status_code, 200)
+        self.assertEqual(resp.json(), [])
+
+        # bad request case
+        params = {
+            "entities": [self.entity.id],
+            "attrs": [{"name": "ref", "attrs": [{"value": "fuga"}]}],
+        }
+        resp = self.client.post(
+            "/entry/api/v2/advanced_search_chain/", json.dumps(params), "application/json"
+        )
+        self.assertEqual(resp.status_code, 400)
+        self.assertEqual(
+            resp.json(),
+            {
+                "non_field_errors": [
+                    {
+                        "code": "AE-121000",
+                        "message": "Invalid condition({'value': 'fuga'}) was specified",
+                    }
+                ]
+            },
         )
 
     def test_entry_history(self):